Bump

Bump.bump

Bump.bump

Bump.bump

Bump.bump

Bump.dbump

Bump.dbump

Bump.dbump

Bump.dbump

Bump.ddbump

Bump.ddbump

Bump.ddbump

Bump.ddbump

Bump.bump — Method

bump(x::AbstractFloat, x0::AbstractFloat, r::AbstractFloat)

Compute a bump function at (x-x0) / r.

Bump.bump — Method

bump(x::AbstractFloat)

Compute a bump function at x.

Bump.bump — Method

bump(x::Vector{AbstractFloat}, x0::Vector{AbstractFloat}, r::Vector{AbstractFloat})

Compute a vector bump function at x centered at x0 scaled by r (element-wise)

Bump.bump — Method

bump(x::Vector{AbstractFloat})

Compute a bump function at the vector x.

Bump.dbump — Method

dbump(x::AbstractFloat, x0::AbstractFloat, r::AbstractFloat)

Compute the first derivative of a bump function at (x-x0) / r.

Bump.dbump — Method

dbump(x::AbstractFloat)

Compute the first derivative of a bump function at x.

Bump.dbump — Method

dbump(x::Vector{AbstractFloat}, x0::Vector{AbstractFloat}, r::Vector{AbstractFloat})

Compute the gradient of a vector bump function at x centered at x0 scaled by r (element-wise)

Bump.dbump — Method

dbump(x::Vector{AbstractFloat})

Compute the first derivative of a bump function at the vector x.

Bump.ddbump — Method

ddbump(x::AbstractFloat, x0::AbstractFloat, r::AbstractFloat)

Compute the second derivative of a bump function at (x-x0) / r.

Bump.ddbump — Method

ddbump(x::AbstractFloat)

Compute the second derivative of a bump function at x.

Bump.ddbump — Method

ddbump(x::Vector{AbstractFloat}, x0::Vector{AbstractFloat}, r::Vector{AbstractFloat})

Compute the Hessian of a vector bump function at x centered at x0 scaled by r (element-wise)

Bump.ddbump — Method

ddbump(x::Vector{AbstractFloat})

Compute the second derivative of a bump function at the vector x.